feat(mcp-page): ui refactor and redesign
