+++ frontend/src/container/MCPServerSettings/CopyIconButton.styles.scss
@@ -1,0 +1,7 @@
+.mcp-copy-btn {
+	color: var(--foreground);
+
+	&:hover {
+		color: var(--bg-robin-400);
+	}
+}
