+++ frontend/src/container/MCPServerSettings/NotCloudFallback/NotCloudFallback.styles.scss
@@ -1,0 +1,28 @@
+.not-cloud-fallback {
+	display: flex;
+	flex-direction: column;
+	gap: var(--spacing-3);
+	padding: var(--padding-6);
+	border: 1px dashed var(--l1-border);
+	border-radius: 8px;
+	background: var(--l1-background);
+	max-width: 40rem;
+
+	&__title {
+		display: flex;
+		align-items: center;
+		gap: var(--spacing-2);
+		font-size: var(--label-large-500-font-size);
+		font-weight: var(--label-large-500-font-weight);
+		color: var(--l1-foreground);
+		margin: 0;
+	}
+
+	&__body {
+		font-size: var(--paragraph-base-400-font-size);
+		font-weight: var(--paragraph-base-400-font-weight);
+		color: var(--foreground);
+		line-height: var(--paragraph-base-400-line-height);
+		margin: 0;
+	}
+}
