+++ frontend/src/container/MCPServerSettings/RegionFallbackCard/RegionFallbackCard.styles.scss
@@ -1,0 +1,35 @@
+.mcp-region-card {
+	display: flex;
+	flex-direction: column;
+	gap: var(--spacing-3);
+	padding: var(--padding-4) var(--padding-5);
+	border: 1px solid var(--l1-border);
+	border-radius: 8px;
+	background: var(--l1-background);
+
+	&__warning {
+		display: flex;
+		align-items: flex-start;
+		gap: var(--spacing-2);
+		padding: var(--padding-3);
+		border: 1px solid var(--bg-amber-500);
+		border-radius: 0.375rem;
+		background: rgba(255, 193, 7, 0.08);
+		font-size: var(--paragraph-small-400-font-size);
+		color: var(--foreground);
+		line-height: var(--paragraph-small-400-line-height);
+
+		svg {
+			flex-shrink: 0;
+			color: var(--bg-amber-500);
+			margin-top: 0.125rem;
+		}
+	}
+
+	&__label {
+		font-size: var(--paragraph-small-400-font-size);
+		font-weight: 500;
+		color: var(--foreground);
+		letter-spacing: 0.01em;
+	}
+}
